Clean up and improve test
- Create test cases for all filetypes
- Change name to align with previous pylint test
- Remove pylint test
- Clean up implementaion
