+++ test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.bash
@@ -1,0 +1,1 @@
+# deprecated: description='expired bash comment' core_version='2.0'
