+++ test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.bat
@@ -1,0 +1,1 @@
+REM deprecated: description='expired bat comment' core_version='2.0'
