+++ test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.cs
@@ -1,0 +1,1 @@
+// deprecated: description='expired cs comment' core_version='2.0'
