+++ test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.fish
@@ -1,0 +1,1 @@
+# deprecated: description='expired fish comment' core_version='2.0'
