+++ test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.ini
@@ -1,0 +1,1 @@
+; deprecated: description='expired ini comment' core_version='2.0'
