+++ test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.ps1
@@ -1,0 +1,1 @@
+# deprecated: description='expired ps1 comment' core_version='2.0'
