+++ test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.py
@@ -1,0 +1,14 @@
+# deprecated: description='valid core thing here', core_version='999.999'
+# deprecated: description='valid python thing here' python_version='999.999'
+# deprecated: description='valid python thing here' python_version='999.999' core_version='999.999'
+# deprecated: description='no version'
+# deprecated: description='invalid key', foo='bar', core_version='999.999'
+# deprecated: description='invalid keys', foo='bar', baz='bax', core_version='999.999'
+# deprecated: 
+# deprecated: description=, core_version='999.999'
+# deprecated: foo, core_version='2.2'
+a = "foo"  # deprecated: description='remove os', core_version='2.0'
+b = "bar"  # deprecated: description='remove os', python_version='2.0'
+# deprecated: description='invalid version' core_version='abc'
+# deprecated: description='unterminated quote
+# deprecated: description='both expired' core_version='2.0' python_version='2.0'
