+++ test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.pyi
@@ -1,0 +1,1 @@
+# deprecated: description='expired pyi comment' core_version='2.0'
