+++ test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+# deprecated: description='expired shell comment' core_version='2.0'
