+++ test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.yaml
@@ -1,0 +1,1 @@
+# deprecated: description='expired yaml comment' core_version='2.0'
