+++ test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases_target.py
@@ -1,0 +1,4 @@
+# deprecated: description='expired for controller but not target' python_version='3.12'
+# deprecated: description='expired for target' python_version='2.0'
+# Note: The minimum versions are hard-coded in runme.sh, so the specified python_versions
+#       will not need updating moving forward.
