+++ test/lib/ansible_test/_util/controller/sanity/pylint/config/collection.cfg
@@ -36,7 +36,6 @@ disable=
     deprecated-argument,  # results vary by Python version
     deprecated-attribute,  # results vary by Python version
     deprecated-class,  # results vary by Python version
-    deprecated-comment,  # custom plugin only used by ansible-core, not collections
     deprecated-decorator,  # results vary by Python version
     deprecated-method,  # results vary by Python version
     deprecated-module,  # results vary by Python version
