+++ test/sanity/code-smell/deprecated-comment.json
