+++ test/sanity/ignore.txt
@@ -268,6 +268,21 @@ test/units/modules/test_apt.py mypy-3.15:name-match
 test/units/modules/test_mount_facts.py mypy-3.15:index
 test/integration/targets/interpreter_discovery_python/library/test_non_python_interpreter.py shebang  # test needs non-standard shebang
 test/integration/targets/inventory_script/bad_shebang shebang  # test needs an invalid shebang
+test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.bash deprecated-comment!skip  # test fixture
+test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.bat deprecated-comment!skip  # test fixture
+test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.cs deprecated-comment!skip  # test fixture
+test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.fish deprecated-comment!skip  # test fixture
+test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.ini deprecated-comment!skip  # test fixture
+test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.ps1 deprecated-comment!skip  # test fixture
+test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.psm1 deprecated-comment!skip  # test fixture
+test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.py deprecated-comment!skip  # test fixture
+test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.py pep8!skip  # test fixture
+test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.py pylint!skip  # test fixture
+test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.pyi deprecated-comment!skip  # test fixture
+test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.sh deprecated-comment!skip  # test fixture
+test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.yaml deprecated-comment!skip  # test fixture
+test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases.yml deprecated-comment!skip  # test fixture
+test/integration/targets/ansible-test-sanity-deprecated-comment/test_cases_target.py deprecated-comment!skip  # test fixture
 test/integration/targets/ansible-test-sanity-pylint/ansible_collections/ns/col/plugins/lookup/deprecated.py pylint!skip  # validated as a collection
 test/integration/targets/ansible-test-sanity-pylint/ansible_collections/ns/col/plugins/action/do_deprecated_stuff.py pylint!skip  # validated as a collection
 test/integration/targets/ansible-test-sanity-pylint/ansible_collections/ns/col/plugins/module_utils/deprecated_utils.py pylint!skip  # validated as a collection
